Disambiguate between communities of same name using lemmy instance
When a user is subscribed to two or more communities of the same on
different instances of lemmy the commands that prompt the user for a
subscribed community need to allow the user to specify which of the
identically named communities the user means. The easiest way to do
this is to include the lemmy instance in the string the user is
prompted for, which is what this commit arranges for.